Debug : filter custom repo 구조 변경 (#121)
