+++ src/main/java/com/adoonge/seedzip/filter/repository/FilterRepositoryImpl.java
@@ -1,0 +1,23 @@
+package com.adoonge.seedzip.filter.repository;
+
+import com.adoonge.seedzip.filter.domain.QFilter;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class FilterRepositoryImpl implements FilterRepositoryCustom {
+
+	private final JPAQueryFactory queryFactory;
+
+	public Long findNextNumber(){
+		QFilter filter = QFilter.filter;
+
+		Long maxNumber = queryFactory.select(filter.filterNum.max())
+			.from(filter)
+			.fetchOne();
+
+		// 필터가 없으면 1부터 시작
+		return maxNumber != null ? maxNumber + 1 : 1;
+	}
+}
